Renamed Window class files & use all srcs files (UI) for exemple
