--- TODO
@@ -1,12 +1,0 @@
-Improve WindowsHandler class
-
-features:
- - profiler for benchmarking and optimization
- - no multithreading (first optimise in monothread and THEN we pump that shit up)
-
-optimizations :
- - Chunk-based storage (32x32x32 ?)			=> obviously
- - Greedy meshing					=> hell yeah
- - Level of Details (LOD)				=> fuck yeah
- - Use OpenCL for meshing & OpenGL for rendering	=> to discuss but probably too convoluted
- - Multithread for LOD generation			=> not usefull and maybe buggy but cool idea
