+++ srcs/events.cpp
@@ -1,6 +1,9 @@
 #include "config.hpp"
 
 // Handle all keyboard & other events
-void	handleEvents(Window window) {
+void	handleEvents(Window &window) {
 	glfwPollEvents();
+
+	if (glfwGetKey(window, GLFW_KEY_ESCAPE) == GLFW_PRESS)
+		glfwSetWindowShouldClose(window, true);
 }
